+++ src/index.ts
@@ -32,6 +32,7 @@ export * from './x/xRollingAverage';
 export * from './x/xRollingMedian';
 export * from './x/xRollingMin';
 export * from './x/xRollingMax';
+export * from './x/xSampling';
 export * from './x/xSubtract';
 export * from './x/xSum';
 export * from './x/xMeanAbsoluteError';
@@ -107,7 +108,7 @@ export * from './matrix/matrixNumericalEncoding';
 export * from './matrix/matrixNumericalDecoding';
 export * from './matrix/matrixApplyNumericalEncoding';
 
-export * from './utils/createSequentialArray';
+export * from './utils/createXArray';
 
 export type DoubleMatrix = DoubleArray[];
 export interface Zone {