+++ src/utils/__tests__/createSequentialArray.test.ts
@@ -1,9 +1,0 @@
-import { createSequentialArray } from '../createSequentialArray';
-
-describe('createSequentialArray', () => {
-  it('normal array', () => {
-    const array = createSequentialArray({ from: 1, to: 4, length: 4 });
-    expect(array).toBeInstanceOf(Float64Array);
-    expect(Array.from(array)).toStrictEqual([1, 2, 3, 4]);
-  });
-});